Add Envato network to equivalent domain sets
Envato is an Alexa top 500 website that asks you to use the same login on many different domains. The login forms are on those domains, not envato.com.

For my use case, account-level domain equivalence rules don’t work because I want to share this with a team, and multiple URLs per login is duplicative because we have multiple Envato logins.

tutsplus.com is another Envato domain that accepts Envato logins, but I believe it also accepts tutsplus.com-specific logins, making it not fully equivalent.