Allow different Backup Dir for DB. Facilitate copy and symlink of/to the latest MSSQL backup. In favour of snapshots/deduplication algorithms, e.g. restic/other automation circumstances.
Facilitate copy and symlink to the of/to the latest MSSQL backup. In favour of snapshots/deduplication algorithms, e.g. restic

minimal comment addtion for the last commit

fix if comparison in the previous commit

fix if comparison ordering in the previous commit

Allow different Backup Dir for DB. Facilitate copy and symlink of/to the latest MSSQL backup. In favour of snapshots/deduplication algorithms, e.g. restic/other automation circumstances.