[AC-1223][AC-1184] Failed Renewals (#3158)
* Added null checks when getting customer metadata

* Added additional logging around paypal payments

* Refactor region validation in StripeController

* Update region retrieval method in StripeController

Refactored the method GetCustomerRegionFromMetadata in StripeController. Previously, it returned null in case of nonexisting region key. Now, it checks all keys with case-insensitive comparison, and if no "region" key is found, it defaults to "US". This was done to handle cases where the region key might not be properly formatted or missing.

* Updated switch expression to be switch statement

* Updated new log to not log user input

* Add handling for 'payment_method.attached' webhook

* Cancelling unpaid premium subscriptions

* Update hardcoded Stripe status strings to constants

* Updated expand string to use snake_case

* Removed unnecessary comments