Prefer compounds for preLaunchTasks
It is not possible to override preLaunchTasks in a compound task.
This leads to issues with concurrency when multiple configurations
depend on a common project.

The suggested workaround is to build unique configurations for each
compound or to create configurations which do not have preLaunchTasks
and then use a compound to launch them.

https://github.com/microsoft/vscode/issues/133468#issuecomment-923211399