PM-10600: Immutable GlobalSettings between test cases, fixture improvements.
The GlobalSettingsFactory no longer provides the same instance of GlobalSettings. If modified, other tests would fail too.
The factory now have a method to create new GlobalSettings objects, which can be modified as it's needed for the test.
This is then used in GlobalSettingsBuilder, which can be used to customize GlobalSettings.