PM-10600: Push notifications sent to installation id.
This enables the Notification Center created global notifications to be sent by an administrator to affected devices of the same server installation. Then all clients connected to any of the server instance of that installation id would receive it.
This is useful for notifying all clients of an installation about upcoming maintenance.
This works both for Self-Hosted, but also for Cloud, assuming an installation id is set.